Split files again for organisation
